decompose upload/save buttons components
