+++ src/client/components/saveButton/saveButton.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './saveButton.css';
-import { saveIcon } from '../../../assets/icons/icons';
+import { saveIcon } from '../../assets/icons/icons';
 
 const SaveButton = () => (
   <button type="button" className="SaveButton">
